tests/functional/requirements.txt
requests==2.26.0
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
elasticsearch==7.9.1
redis==4.1.1
django-environ==0.8.1
